+++ src/game/tests.rs
@@ -572,6 +572,7 @@ mod vintage_cube_tireless_tracker;
 mod vintage_cube_tishanas_tidebinder;
 mod vintage_cube_titania;
 mod vintage_cube_touch_spirit_realm;
+mod vintage_cube_traveling_chocobo;
 mod vintage_cube_troll_of_khazad_dum;
 mod vintage_cube_twisted_landscape;
 mod vintage_cube_ugin_eye_of_the_storms;